+++ main.tf
@@ -212,6 +212,7 @@ resource "kubernetes_secret" "this" {
       {
         "app.kubernetes.io/instance" = var.secret_name
       },
+      local.labels,
       var.labels,
       var.secret_labels
     )