+++ outputs.tf
@@ -13,17 +13,17 @@ output "grafana_dashboards" {
 }
 
 output "image_name" {
-  description = "Name of the docker image used for the cloudwatch-exporter container."
+  description = "Name of the docker image used for the aws-health-status container."
   value       = var.enabled ? var.image_name : ""
 }
 
 output "image_version" {
-  description = "Tag of the docker image used for the cloudwatch-exporter container."
+  description = "Tag of the docker image used for the aws-health-status container."
   value       = var.enabled ? var.image_version : ""
 }
 
 output "image_pull_policy" {
-  description = "Image pull policy defined on the cloudwatch-exporter container."
+  description = "Image pull policy defined on the aws-health-status container."
   value       = var.enabled ? var.image_pull_policy : ""
 }