+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 0.2.0 / 2020-02-27
 ==================
 
+  * tech: Use KinD for testing
   * feat: Add grafana_dashboards output
   * tech: Update pre-commit configuration