+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 0.3.0 / 2020-02-27
 ==================
 
+  * fix: Add local labels on secret
   * fix: Add local for service port
   * tech: Add outputs on example
   * feat: Add deployment template variables, outputs